python_renderer/requirements.txt
Flask
pycairo
uharfbuzz
PyGObject
